requirements.txt
uvicorn==0.11.3
fastapi==0.52.0
mongoengine==0.19.1
pydantic==1.4
bcrypt==3.1.7
PyJWT==1.7.1
python-dotenv==0.12.0
PyJWT==1.7.1